requirements.txt
Genshi==0.6
M2Crypto==0.20.2
Thrift==0.2.0
Twisted==10.2.0
amqplib==0.6.1
anyjson==0.2.5
carrot==0.10.14-txamqp
gviz-api.py==1.7.0
httplib2==0.6.0
msgpack-python==015final
simplejson==2.1.2
telephus==0.7-beta3.3
txAMQP==0.3
zope.interface==3.6.1
ionproto==0.3.12
protobuf==2.3.0-p1
.
